use DocumentArchive component for single and bulk archiving
