+++ shared/i18n/locales/en_US/translation.json
@@ -97,8 +97,6 @@
   "Move to collection": "Move to collection",
   "Move {{ documentType }}": "Move {{ documentType }}",
   "Are you sure you want to archive this document?": "Are you sure you want to archive this document?",
-  "Document archived": "Document archived",
-  "Archiving this document will remove it from the collection and search results.": "Archiving this document will remove it from the collection and search results.",
   "{{ documentName }} restored": "{{ documentName }} restored",
   "Choose a collection": "Choose a collection",
   "Delete {{ documentName }}": "Delete {{ documentName }}",
@@ -183,11 +181,6 @@
   "currently viewing": "currently viewing",
   "previously edited": "previously edited",
   "You": "You",
-  "{{ count }} documents archived": "{{ count }} documents archived",
-  "{{ count }} documents archived_plural": "{{ count }} documents archived",
-  "{{ successCount }} archived, {{ errorCount }} failed": "{{ successCount }} archived, {{ errorCount }} failed",
-  "Are you sure you want to archive <em>{{ count }} documents</em>? They will be removed from collections and search results.": "Are you sure you want to archive <em>{{ count }} documents</em>? They will be removed from collections and search results.",
-  "Are you sure you want to archive <em>{{ count }} documents</em>? They will be removed from collections and search results._plural": "Are you sure you want to archive <em>{{ count }} documents</em>? They will be removed from collections and search results.",
   "Archive {{ count }} documents": "Archive {{ count }} documents",
   "Archive {{ count }} documents_plural": "Archive {{ count }} documents",
   "Delete {{ count }} documents": "Delete {{ count }} documents",
@@ -481,6 +474,7 @@
   "{{ documentName }} cannot be moved within {{ parentDocumentName }}": "{{ documentName }} cannot be moved within {{ parentDocumentName }}",
   "You can't reorder documents in an alphabetically sorted collection": "You can't reorder documents in an alphabetically sorted collection",
   "The {{ documentName }} cannot be moved here": "The {{ documentName }} cannot be moved here",
+  "Document archived": "Document archived",
   "Return to App": "Back to App",
   "Installation": "Installation",
   "Unstar document": "Unstar document",
@@ -791,6 +785,12 @@
   "Observing {{ userName }}": "Observing {{ userName }}",
   "Backlinks": "Backlinks",
   "This document is large which may affect performance": "This document is large which may affect performance",
+  "{{ count }} documents archived": "{{ count }} documents archived",
+  "{{ count }} documents archived_plural": "{{ count }} documents archived",
+  "{{ successCount }} archived, {{ errorCount }} failed": "{{ successCount }} archived, {{ errorCount }} failed",
+  "Are you sure you want to archive <em>{{ count }} documents</em>? They will be removed from collections and search results.": "Are you sure you want to archive <em>{{ count }} documents</em>? They will be removed from collections and search results.",
+  "Are you sure you want to archive <em>{{ count }} documents</em>? They will be removed from collections and search results._plural": "Are you sure you want to archive <em>{{ count }} documents</em>? They will be removed from collections and search results.",
+  "Archiving this document will remove it from the collection and search results.": "Archiving this document will remove it from the collection and search results.",
   "{{ successCount }} deleted, {{ errorCount }} failed": "{{ successCount }} deleted, {{ errorCount }} failed",
   "{{ count }} documents deleted": "{{ count }} documents deleted",
   "{{ count }} documents deleted_plural": "{{ count }} documents deleted",
